requirements.txt
# Core ML
torch>=2.1.0
transformers>=4.36.0
sentence-transformers>=2.2.2
accelerate>=0.24.0

# NLP utilities
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0

# Data handling
datasets>=2.14.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
jsonlines>=3.1.0

# Optional: OpenAI for evaluation only
openai>=1.0.0
